Compat with PaddlePaddle

+++ python/cudnn/deepseek_sparse_attention/indexer_backward/indexer_backward_sm90.py
@@ -1736,7 +1736,7 @@ def _build_cute_dsl_kernel(
                 current_stream=current_stream,
             )
             with _torch_stream_context(current_stream):
-                dIndexK.copy_(dIndexK_f32)
+                dIndexK.copy_(dIndexK_f32.astype(dIndexK.dtype))
 
     _run.score_grad = partial(_score_grad_inplace, index_is_log=score_input_is_log)
     _run.gemm_only = _run_gemm_only
